Fixed per reviewers' comments. Changed the package name to be more specific, added package-info.java, added to more build targets.
