+++ default.properties
@@ -170,6 +170,7 @@ plugins.index=\
    org.apache.nutch.indexer.basic*:\
    org.apache.nutch.indexer.feed*:\
    org.apache.nutch.indexer.geoip*:\
+   org.apache.nutch.indexer.jexl*:\
    org.apache.nutch.indexer.filter*:\
    org.apache.nutch.indexer.links*:\
    org.apache.nutch.indexer.metadata*:\
